// employee/src/key-value/ts/enums/code.enum.ts
export enum CodeEnum {
  CURRENCY = 'CURRENCY',
  COUNTRY = 'COUNTRY',
  NATIONALITY = 'NATIONALITY',
  IMMIGRATION_STATUS = 'IMMIGRATION_STATUS'
}

export enum CODE_RELATIONSHIP_PERMISSION {
  CREATE_RELATIONSHIP = 'CREATE_RELATIONSHIP',
  UPDATE_RELATIONSHIP = 'UPDATE_RELATIONSHIP',
  READ_RELATIONSHIP = 'READ_RELATIONSHIP',
  DELETE_RELATIONSHIP = 'DELETE_RELATIONSHIP'
}

export enum CODE_DOCUMENT_TYPE_PERMISSION {
  CREATE_DOCUMENT_TYPE = 'CREATE_DOCUMENT_TYPE',
  UPDATE_DOCUMENT_TYPE = 'UPDATE_DOCUMENT_TYPE',
  READ_DOCUMENT_TYPE = 'READ_DOCUMENT_TYPE',
  DELETE_DOCUMENT_TYPE = 'DELETE_DOCUMENT_TYPE'
}

export enum CODE_EDUCATION_TYPE_PERMISSION {
  CREATE_EDUCATION_TYPE = 'CREATE_EDUCATION_TYPE',
  UPDATE_EDUCATION_TYPE = 'UPDATE_EDUCATION_TYPE',
  READ_EDUCATION_TYPE = 'READ_EDUCATION_TYPE',
  DELETE_EDUCATION_TYPE = 'DELETE_EDUCATION_TYPE'
}

export enum CODE_TRAINING_PERMISSION {
  CREATE_TRAINING = 'CREATE_TRAINING',
  UPDATE_TRAINING = 'UPDATE_TRAINING',
  READ_TRAINING = 'READ_TRAINING',
  DELETE_TRAINING = 'DELETE_TRAINING'
}

export enum CODE_SKILL_PERMISSION {
  CREATE_SKILL = 'CREATE_SKILL',
  UPDATE_SKILL = 'UPDATE_SKILL',
  READ_SKILL = 'READ_SKILL',
  DELETE_SKILL = 'DELETE_SKILL'
}

export enum CODE_RESIGNATION_TYPE {
  CREATE_RESIGNATION_TYPE = 'CREATE_RESIGNATION_TYPE',
  UPDATE_RESIGNATION_TYPE = 'UPDATE_RESIGNATION_TYPE',
  READ_RESIGNATION_TYPE = 'READ_RESIGNATION_TYPE',
  DELETE_RESIGNATION_TYPE = 'DELETE_RESIGNATION_TYPE'
}

export enum CODE_WARNING_TYPE_PERMISSION {
  CREATE_WARNING_TYPE = 'CREATE_WARNING_TYPE',
  UPDATE_WARNING_TYPE = 'UPDATE_WARNING_TYPE',
  READ_WARNING_TYPE = 'READ_WARNING_TYPE',
  DELETE_WARNING_TYPE = 'DELETE_WARNING_TYPE'
}

export enum CODE_COMPANY_STRUCTURE_COMPONENT {
  CREATE_COMPANY_STRUCTURE_COMPONENT = 'CREATE_COMPANY_STRUCTURE_COMPONENT_COLOR',
  UPDATE_COMPANY_STRUCTURE_COMPONENT = 'UPDATE_COMPANY_STRUCTURE_COMPONENT_COLOR',
  READ_COMPANY_STRUCTURE_COMPONENT = 'READ_COMPANY_STRUCTURE_COMPONENT_COLOR',
  DELETE_COMPANY_STRUCTURE_COMPONENT = 'DELETE_COMPANY_STRUCTURE_COMPONENT_COLOR'
}