// employee/src/key-value/ts/enums/permission.enum.ts
import * as PERMISSION from '../../../shared-resources/ts/enum/permission';

export const CODE_VALUE = {
  GET_ALL_VALUE_BY_KEY: [
    PERMISSION.CODE_VALUE_PERMISSION.READ_CODE_VALUE,
    PERMISSION.SPECIAL_PERMISSION.ALL_FUNCTION,
    PERMISSION.SPECIAL_PERMISSION.READ_ALL_FUNCTION
  ]
};

export enum CodeTypesEnum {
  RELATIONSHIP = 'RELATIONSHIP',
  DOCUMENT_TYPE = 'DOCUMENT_TYPE',
  EDUCATION_TYPE = 'EDUCATION_TYPE',
  TRAINING = 'TRAINING',
  SKILL = 'SKILL',
  RESIGNATION_TYPE = 'RESIGNATION_TYPE',
  WARNING_TYPE = 'WARNING_TYPE',
  MOVEMENT = 'MOVEMENT',
  NATIONALITY = 'NATIONALITY',
  LANGUAGE = 'LANGUAGE',
  CURRENCY = 'CURRENCY',
  COUNTRY = 'COUNTRY',
  COMPANY_STRUCTURE_COMPONENT_COLOR = 'COMPANY_STRUCTURE_COMPONENT_COLOR',
  GENDER = 'GENDER',
  MARITAL_STATUS = 'MARITAL_STATUS'
}